MTV-2511: Refactor the providers YAML tab
Reference: https://issues.redhat.com/browse/MTV-2511
Subtask of: https://issues.redhat.com/browse/MTV-1982

 - Refactor the YAML tab of the providers details pages.
 - Remove part of the old pre-refactored code.
