+++ src/providers/details/OpenshiftProviderDetailsPage.tsx
@@ -3,7 +3,6 @@ import { ModalHOC } from 'src/modules/Providers/modals/ModalHOC/ModalHOC';
 import { ProviderPageHeadings } from 'src/modules/Providers/views/details/components/ProviderPageHeadings';
 import { ProviderCredentialsTabPage } from 'src/modules/Providers/views/details/tabs/Credentials/ProviderCredentials';
 import { ProviderVirtualMachinesTabPage } from 'src/modules/Providers/views/details/tabs/VirtualMachines/ProviderVirtualMachines';
-import { ProviderYAMLTabPage } from 'src/modules/Providers/views/details/tabs/YAML/ProviderYAML';
 import { useForkliftTranslation } from 'src/utils/i18n';
 
 import type { V1beta1Provider } from '@kubev2v/types';
@@ -11,6 +10,7 @@ import { HorizontalNav, type NavPage } from '@openshift-console/dynamic-plugin-s
 
 import ProviderDetailsTabPage from './tabs/Details/ProviderDetailsTabPage';
 import ProviderNetworksTabPage from './tabs/Networks/ProviderNetworksTabPage';
+import ProviderYAMLTabPage from './tabs/YAML/ProviderYAMLTabPage';
 
 const OpenshiftProviderDetailsPage: FC<{
   name: string;
@@ -26,7 +26,7 @@ const OpenshiftProviderDetailsPage: FC<{
       name: t('Details'),
     },
     {
-      component: () => <ProviderYAMLTabPage name={name} namespace={namespace} />,
+      component: () => <ProviderYAMLTabPage provider={provider} />,
       href: 'yaml',
       name: t('YAML'),
     },
