+++ src/providers/details/VSphereProviderDetailsPage.tsx
@@ -3,13 +3,13 @@ import { ProviderPageHeadings } from 'src/modules/Providers/views/details/compon
 import { ProviderCredentialsTabPage } from 'src/modules/Providers/views/details/tabs/Credentials/ProviderCredentials';
 import { ProviderHostsTabPage } from 'src/modules/Providers/views/details/tabs/Hosts/ProviderHosts';
 import { ProviderVirtualMachinesTabPage } from 'src/modules/Providers/views/details/tabs/VirtualMachines/ProviderVirtualMachines';
-import { ProviderYAMLTabPage } from 'src/modules/Providers/views/details/tabs/YAML/ProviderYAML';
 import { useForkliftTranslation } from 'src/utils/i18n';
 
 import type { V1beta1Provider } from '@kubev2v/types';
 import { HorizontalNav, type NavPage } from '@openshift-console/dynamic-plugin-sdk';
 
 import ProviderDetailsTabPage from './tabs/Details/ProviderDetailsTabPage';
+import ProviderYAMLTabPage from './tabs/YAML/ProviderYAMLTabPage';
 
 const VSphereProviderDetailsPage: FC<{
   name: string;
@@ -25,7 +25,7 @@ const VSphereProviderDetailsPage: FC<{
       name: t('Details'),
     },
     {
-      component: () => <ProviderYAMLTabPage name={name} namespace={namespace} />,
+      component: () => <ProviderYAMLTabPage provider={provider} />,
       href: 'yaml',
       name: t('YAML'),
     },
